test(research): extract weak slab fixtures
